Actually, no, I take that back. Instead, force .npc add and .gobj add to play ball, stupid commands that they are.
This means that spawnGroupData is always a valid value, never nullptr, and we no longer need to check it anywhere (so removed those checks everywhere).

Also, get rid of some memory usage by not storing SYSTEM groups in the lookup multimap, since we only use the multimap for manual spawning/despawning (which system groups can't do).